+++ internal/domain/event/event_type.go
@@ -1,0 +1,5 @@
+package event
+
+// EventType - enumeration for event type.
+// Value Object
+type EventType int